requirements.txt
bcrypt            ==4.1.2
colorama          ==0.4.6
Flask             ==3.0.0
Flask-Login       ==0.6.3
Flask-SQLAlchemy  ==3.1.1
python-dotenv     ==1.0.0
SQLAlchemy        ==2.0.25
Werkzeug          ==3.0.1
gunicorn          ==20.1.0
psycopg2-binary   ==2.9.9
Flask-Migrate     ==4.0.5